requirements.txt
nats-py
psycopg2-binary
pandas
ccxt
